requirements.txt
# Flask and extensions

Flask==2.3.3
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0

# HTTP requests

requests==2.31.0

# Utilities

Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.2

# Security

WTForms==3.0.1

# API integrations

exa-py==1.0.6

# Development tools (optional)

pytest==7.4.0
pytest-flask==1.2.0
coverage==7.3.1
